Create EntryPlaceholders class
