+++ src/gui/entry/AutoTypeAssociationsModel.cpp
@@ -1,4 +1,5 @@
 /*
+ *  Copyright (C) 2026 KeePassXC Team <team@keepassxc.org>
  *  Copyright (C) 2012 Felix Geyer <debfx@fobos.de>
  *
  *  This program is free software: you can redistribute it and/or modify
@@ -96,7 +97,7 @@ QVariant AutoTypeAssociationsModel::data(const QModelIndex& index, int role) con
                 return tr("(empty)");
             }
             if (m_entry) {
-                window = m_entry->maskPasswordPlaceholders(window);
+                window = EntryPlaceholders::maskPasswordPlaceholders(window);
                 window = m_entry->resolveMultiplePlaceholders(window);
             }
             return window;
